Lazy update
Everything is out of order here, but enough such that I can implement
instructions one at a time. This is not, repeat, not at all how the code
will look; think of it as humble beginnings. Lots of work to do, that I
don't want to lose on account of my SSDs at a critical juncture.

Signed-off-by: dgz <dgz@dgz.me>