Implement cartridge support
Only ROM ONLY and MBC1 MBCs are supported at this time. MBC1 is used by
blargg's CPU tests, which is what we're going to target first. Some
logger changes were needed (LOG_ERR didn't exist!), so that was added
here, too.

comp.h is used to abstract gcc/clang specifics that we use into simpler
forms to use. In this case, it is used for the likely/unlikely macros
in agoge_core_cart_set.

Signed-off-by: dgz <dgz@dgz.me>