Add bit operation macros
These are preferred, because if you're staring at a screen for a
while, you may not see you've missed a & for the &= operation.

But, these are also to be used situationally. Do not try and
formulate some crazy complicated expression using these.

I've added singular and plural forms of these macros.

Signed-off-by: dgz <dgz@dgz.me>