Increase cohesion of agoge_core_ctx
It does not make sense to pass pointers in an OOP like way to various
subcomponents that need access to e.g., the logger, and soon to be the
scheduler. Everything should be encapsulated in an agoge_core_ctx and
handled through the context, and this avoids a problem where for example
the CPU has a pointer to the bus and the logger, but the bus also has a
pointer to the logger... you can see how quickly this might become a
mess.

Work is still not done in this area - additional clean up is coming.

I do not want to proceed further until we have proper disassembly
tracing and cleanups are complete; the next step is the scheduler and
timers to allow for instr-timing.gb to execute and pass.

Signed-off-by: dgz <dgz@dgz.me>